requirements.txt
python-aoc
python-dotenv